pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sectionminer"
version = "0.1.0"
description = "Extract sections and subsections from academic PDFs"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "SectionMiner Contributors" }]
dependencies = [
  "pymupdf",
  "langchain",
  "langchain-openai",
  "langchain-text-splitters",
  "langchain-community",
  "python-decouple",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["sectionminer*"]

